+++ src/components/blog/blog.style.js
@@ -6,7 +6,11 @@ const styles = theme => ({
   },
   'primary-card': {
     background: theme.palette.primary.light,
-  }
+  },
+  headerTheme: {
+    backgroundColor: theme.palette.primary.main,
+    color: theme.palette.text.primary
+  },
 });
 
 export default styles;